Feat: Add NoticeBoard data unit class
